Add flag --json_frequencies_array
VEP's default JSON format gives allele frequencies for co-located
variants as a hash/dictionary keyed on allele. This doesn't play
well with parsers that expect a fixed schema as the key names
can't be predicted.

This flag re-formats the frequencies hash as a list of hashrefs
with an 'allele' key added.